pyproject.toml
[tool.poetry]
name = "instachatico"
version = "0.1.0"
description = ""
authors = ["Mikhail Denisov <denisovmvtmb@yandex.ru>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["standard"], version = "^0.117.1"}
uvicorn = {extras = ["standard"], version = "^0.36.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.43"}
python-dotenv = "^1.1.1"
alembic = "^1.16.5"
pydantic-settings = "^2.10.1"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.9"
celery = "^5.5.3"
redis = "^6.4.0"
langchain-community = "^0.3.29"
aiohttp = "^3.9.0"
openai-agents = "^0.3.2"
pgvector = "^0.3.6"
dependency-injector = "^4.41.0"

# Document intelligence
boto3 = "^1.34.0"
python-magic = "^0.4.27"
pdfplumber = "^0.11.0"
python-docx = "^1.0.0"
pandas = "^2.0.0"
openpyxl = "^3.1.0"
aiolimiter = "^1.2.1"
PyJWT = ">=2.10.1,<3.0.0"

[tool.poetry.group.dev.dependencies]
black = "^24.0.0"
pytest = "^8.4.2"
pytest-asyncio = "^0.24.0"
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
httpx = "^0.27.0"
faker = "^30.0.0"
aiosqlite = "^0.21.0"
fakeredis = "^2.23.0"
pytest-html = "^4.1.1"

[tool.black]
line-length = 120
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
